test: convert emulator-independent integration tests to pure unit + consolidate duplicates [wave 2a]
Convert lifecycle + data-helpers to pure MemoryDataService-backed tests,
consolidate gallery-view→social-features and title-persistence→layout-saving.
Keep hybrid-integration emulator-backed (Memory ignores batchSearchFn).
