+++ recipe-lanes/package.json
@@ -13,7 +13,7 @@
     "typecheck": "tsc --noEmit",
     "test": "npm run test:unit && npm run test:e2e",
     "test:unit": "npm run test:unit:pure && npm run test:unit:integration",
-    "test:unit:pure": "npx env-cmd -f .env.test node --import tsx --test tests/data.test.ts tests/gallery-view.test.ts tests/graph-logic.test.ts tests/graph.test.ts tests/hybrid-plumbing.test.ts tests/icon-pipeline.test.ts tests/icon-search.test.ts tests/icon-shortlist.test.ts tests/image-processing.test.ts tests/layout-saving.test.ts tests/model-utils.test.ts tests/optimistic-flow.test.ts tests/parser.test.ts tests/patch.test.ts tests/recipe-store.test.ts tests/save-copy.test.ts tests/shortlist-delta.test.ts tests/social-features.test.ts tests/timeline-layout.test.ts tests/timeline-save.test.ts tests/title-persistence.test.ts tests/undo-complex.test.ts tests/verify-production-logic.test.ts",
+    "test:unit:pure": "npx env-cmd -f .env.test node --import tsx --test tests/data.test.ts tests/data-helpers.test.ts tests/graph-logic.test.ts tests/graph.test.ts tests/hybrid-plumbing.test.ts tests/icon-pipeline.test.ts tests/icon-search.test.ts tests/icon-shortlist.test.ts tests/image-processing.test.ts tests/layout-saving.test.ts tests/lifecycle.test.ts tests/model-utils.test.ts tests/optimistic-flow.test.ts tests/parser.test.ts tests/patch.test.ts tests/recipe-store.test.ts tests/save-copy.test.ts tests/shortlist-delta.test.ts tests/social-features.test.ts tests/timeline-layout.test.ts tests/timeline-save.test.ts tests/undo-complex.test.ts tests/verify-production-logic.test.ts",
     "test:unit:integration": "./scripts/test-unit-integration.sh",
     "test:one": "npx env-cmd -f .env.test node --import tsx --test",
     "test:watch": "node --import tsx --test --watch tests/*.test.ts",
