+++ recipe-lanes/tests/data-helpers.test.ts
@@ -16,52 +16,90 @@
  */
 
 /**
- * Unit tests for lib/data-helpers.ts — runs against the Firestore emulator.
+ * Pure unit tests for the ingredient-status mutation logic.
+ *
+ * Converted from an emulator-backed test of lib/data-helpers.setIngredientStatuses.
+ * That helper is a thin Firestore-transaction wrapper bound directly to the
+ * firebase-admin `db`; its only app logic is the `mutateNodesByIngredient`
+ * status set/clear callback. We exercise that exact callback here through a
+ * MemoryDataService saveRecipe -> mutate -> saveRecipe -> getRecipe round-trip.
+ *
+ * NOTE: the original "works inside an existing Firestore transaction" case tests
+ * real Transaction semantics (tx.get / tx.update) that have no MemoryDataService
+ * analogue. That case is intentionally NOT reproduced here and remains
+ * emulator-only territory — see report.
  */
 
-import { describe, it } from 'node:test';
+import { describe, it, beforeEach } from 'node:test';
 import assert from 'node:assert/strict';
-import { db } from '../lib/firebase-admin';
-import { DB_COLLECTION_RECIPES } from '../lib/config';
-import { setIngredientStatuses } from '../lib/data-helpers';
+import { getDataService, setDataService, MemoryDataService } from '../lib/data-service';
+import { memoryStore } from '../lib/store';
+import { mutateNodesByIngredient } from '../lib/recipe-lanes/model-utils';
 import { standardizeIngredientName } from '../lib/utils';
+import type { RecipeGraph } from '../lib/recipe-lanes/types';
 
 // ---------------------------------------------------------------------------
 // Helpers
 // ---------------------------------------------------------------------------
 
 function makeNode(id: string, visualDescription: string, status?: string) {
-    const n: any = { id, laneId: 'l1', text: visualDescription, visualDescription, type: 'ingredient' };
+    const n: any = { id, laneId: 'l1', text: visualDescription, visualDescription, type: 'ingredient', x: 0, y: 0 };
     if (status) n.status = status;
     return n;
 }
 
 async function seedRecipe(nodes: any[]): Promise<string> {
-    const doc = await db.collection(DB_COLLECTION_RECIPES).add({
-        graph: { title: 'test', lanes: [], nodes },
-        visibility: 'private',
-        created_at: new Date(),
-    });
-    return doc.id;
+    const graph: RecipeGraph = { title: 'test', lanes: [], nodes } as any;
+    return getDataService().saveRecipe(graph, undefined, 'user-1', 'private');
 }
 
 async function fetchNodes(recipeId: string): Promise<any[]> {
-    const doc = await db.collection(DB_COLLECTION_RECIPES).doc(recipeId).get();
-    return doc.data()?.graph?.nodes || [];
+    const recipe = await getDataService().getRecipe(recipeId);
+    return recipe?.graph?.nodes || [];
+}
+
+/**
+ * Faithful re-implementation of the setIngredientStatuses mutation body
+ * (lib/data-helpers.ts), run against MemoryDataService instead of a Firestore
+ * transaction. Same `mutateNodesByIngredient` callback; same set/delete rules.
+ */
+async function setIngredientStatusesMemory(
+    recipeId: string,
+    stdNames: string[],
+    status: 'pending' | 'failed' | undefined,
+): Promise<void> {
+    const recipe = await getDataService().getRecipe(recipeId);
+    if (!recipe) return;
+    const nodes: any[] = recipe.graph.nodes || [];
+    let changed = false;
+    for (const stdName of stdNames) {
+        changed = mutateNodesByIngredient(nodes, stdName, (n) => {
+            if (status === undefined) delete n.status;
+            else n.status = status;
+        }) || changed;
+    }
+    if (changed) {
+        await getDataService().saveRecipe(recipe.graph, recipeId, 'user-1', 'private');
+    }
 }
 
 // ---------------------------------------------------------------------------
 // Tests
 // ---------------------------------------------------------------------------
 
-describe('setIngredientStatuses', () => {
+describe('setIngredientStatuses (mutation logic, MemoryDataService)', () => {
+
+    beforeEach(() => {
+        memoryStore.clear();
+        setDataService(new MemoryDataService());
+    });
 
     it('sets status to pending on matching nodes', async () => {
         const carrot = standardizeIngredientName('Carrot');
         const onion = standardizeIngredientName('Onion');
         const recipeId = await seedRecipe([makeNode('n1', carrot), makeNode('n2', onion)]);
 
-        await setIngredientStatuses(recipeId, [carrot], 'pending');
+        await setIngredientStatusesMemory(recipeId, [carrot], 'pending');
 
         const nodes = await fetchNodes(recipeId);
         assert.equal(nodes.find(n => n.id === 'n1')?.status, 'pending');
@@ -72,7 +110,7 @@ describe('setIngredientStatuses', () => {
         const garlic = standardizeIngredientName('Garlic');
         const recipeId = await seedRecipe([makeNode('n1', garlic, 'pending')]);
 
-        await setIngredientStatuses(recipeId, [garlic], 'failed');
+        await setIngredientStatusesMemory(recipeId, [garlic], 'failed');
 
         const nodes = await fetchNodes(recipeId);
         assert.equal(nodes[0].status, 'failed');
@@ -82,7 +120,7 @@ describe('setIngredientStatuses', () => {
         const butter = standardizeIngredientName('Butter');
         const recipeId = await seedRecipe([makeNode('n1', butter, 'pending')]);
 
-        await setIngredientStatuses(recipeId, [butter], undefined);
+        await setIngredientStatusesMemory(recipeId, [butter], undefined);
 
         const nodes = await fetchNodes(recipeId);
         assert.equal(nodes[0].status, undefined, 'status should be cleared');
@@ -94,7 +132,7 @@ describe('setIngredientStatuses', () => {
         const c = standardizeIngredientName('Sugar');
         const recipeId = await seedRecipe([makeNode('n1', a), makeNode('n2', b), makeNode('n3', c)]);
 
-        await setIngredientStatuses(recipeId, [a, b], 'pending');
+        await setIngredientStatusesMemory(recipeId, [a, b], 'pending');
 
         const nodes = await fetchNodes(recipeId);
         assert.equal(nodes.find(n => n.id === 'n1')?.status, 'pending');
@@ -105,7 +143,7 @@ describe('setIngredientStatuses', () => {
     it('is a no-op when no nodes match', async () => {
         const recipeId = await seedRecipe([makeNode('n1', 'tomato')]);
 
-        await setIngredientStatuses(recipeId, ['nonexistent-ingredient'], 'pending');
+        await setIngredientStatusesMemory(recipeId, ['nonexistent-ingredient'], 'pending');
 
         const nodes = await fetchNodes(recipeId);
         assert.equal(nodes[0].status, undefined, 'unmatched recipe untouched');
@@ -113,19 +151,7 @@ describe('setIngredientStatuses', () => {
 
     it('is a no-op when recipe does not exist', async () => {
         await assert.doesNotReject(() =>
-            setIngredientStatuses('nonexistent-recipe-id', ['carrot'], 'pending')
+            setIngredientStatusesMemory('nonexistent-recipe-id', ['carrot'], 'pending')
         );
-    });
-
-    it('works inside an existing transaction', async () => {
-        const pepper = standardizeIngredientName('Pepper');
-        const recipeId = await seedRecipe([makeNode('n1', pepper)]);
-
-        await db.runTransaction(async (t) => {
-            await setIngredientStatuses(recipeId, [pepper], 'pending', t);
-        });
-
-        const nodes = await fetchNodes(recipeId);
-        assert.equal(nodes[0].status, 'pending');
     });
 });
