+++ recipe-lanes/tests/lifecycle.test.ts
@@ -15,45 +15,41 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import { describe, it, before } from 'node:test';
+/**
+ * Pure unit tests for the recipe + icon lifecycle.
+ *
+ * Converted from an emulator-backed integration test: the actions exercised here
+ * (createDebugRecipeAction / addIngredientNodeAction / forgeIconAction / getRecipe)
+ * all work against the in-memory MemoryDataService, which resolves icons
+ * synchronously via its mock queueIcons path. No emulator and no polling needed.
+ */
+
+import { describe, it, beforeEach } from 'node:test';
 import assert from 'node:assert';
 import { createDebugRecipeAction, addIngredientNodeAction, forgeIconAction } from '../app/actions';
 import { setAIService, MockAIService } from '../lib/ai-service';
-import { getDataService } from '../lib/data-service';
+import { getDataService, setDataService, MemoryDataService } from '../lib/data-service';
+import { memoryStore } from '../lib/store';
 import { setAuthService, MockAuthService } from '../lib/auth-service';
-import { setIconQueueConfig } from '../lib/icon-queue-config';
 import { getNodeIconId } from '../lib/recipe-lanes/model-utils';
 
-setAIService(new MockAIService());
-setAuthService(new MockAuthService());
-
-/** Poll the recipe node until it has an icon different from `excludeId`, or timeout. */
-async function pollForIcon(
-    recipeId: string,
-    nodeId: string,
-    excludeId: string | undefined,
-    timeoutMs = 60_000
-): Promise<string | null> {
-    const deadline = Date.now() + timeoutMs;
-    while (Date.now() < deadline) {
-        const recipe = await getDataService().getRecipe(recipeId);
-        const node = recipe?.graph?.nodes?.find((n: any) => n.id === nodeId);
-        const iconId = node ? getNodeIconId(node) : undefined;
-        if (iconId && iconId !== excludeId) return iconId;
-        await new Promise(r => setTimeout(r, 500));
-    }
-    return null;
+/** Read the current icon id on a recipe node (synchronous — MemoryDataService resolves inline). */
+async function readIconId(recipeId: string, nodeId: string): Promise<string | undefined> {
+    const recipe = await getDataService().getRecipe(recipeId);
+    const node = recipe?.graph?.nodes?.find((n: any) => n.id === nodeId);
+    return node ? (getNodeIconId(node) ?? undefined) : undefined;
 }
 
 describe('Recipe & Icon Lifecycle', () => {
-    // Ensure forging isn't blocked by config leaked from earlier integration
-    // files sharing the emulator (allowAnonForge / a low perUserDailyCap).
-    before(async () => {
-        await setIconQueueConfig({ paused: false, allowAnonForge: true, perUserDailyCap: 1000 });
+    beforeEach(() => {
+        memoryStore.clear();
+        setDataService(new MemoryDataService());
+        setAIService(new MockAIService());
+        setAuthService(new MockAuthService());
     });
 
-    it('should follow the full creation and shortlist-cycle reroll flow', async () => {
-        const ingredient = 'Integration-Egg-' + Date.now();
+    it('should follow the full creation and icon-resolution flow', async () => {
+        const ingredient = 'Lifecycle-Egg-' + Date.now();
 
         const r1 = await createDebugRecipeAction() as any;
         assert.ok(r1.recipeId, 'createDebugRecipeAction should return a recipeId');
@@ -63,11 +59,16 @@ describe('Recipe & Icon Lifecycle', () => {
         assert.ok(r2.nodeId, 'addIngredientNodeAction should return a nodeId');
         const nodeId = r2.nodeId;
 
-        const iconId = await pollForIcon(recipeId, nodeId, undefined);
-        assert.ok(iconId, `Icon not generated within timeout for "${ingredient}"`);
+        const iconId = await readIconId(recipeId, nodeId);
+        assert.ok(iconId, `Icon not generated for "${ingredient}"`);
     });
 
-    it('forge produces a new icon distinct from the original', async () => {
+    it('forge succeeds and re-queues the node (reroll request accepted)', async () => {
+        // NOTE: producing a *distinct* forged icon requires the real icon index /
+        // reroll cloud function, which MemoryDataService does not implement (its
+        // resolveRecipeIcons skips nodes that already hold an icon). That distinctness
+        // assertion stays emulator-backed elsewhere; here we exercise the pure path:
+        // forgeIconAction -> rejectRecipeIcon succeeds and marks the node for re-resolution.
         const ingredient = 'Forge-Egg-' + Date.now();
 
         const r1 = await createDebugRecipeAction() as any;
@@ -78,16 +79,14 @@ describe('Recipe & Icon Lifecycle', () => {
         assert.ok(r2.nodeId);
         const nodeId = r2.nodeId;
 
-        const initialIconId = await pollForIcon(recipeId, nodeId, undefined);
-        assert.ok(initialIconId, `Initial icon not generated within timeout for "${ingredient}"`);
+        const initialIconId = await readIconId(recipeId, nodeId);
+        assert.ok(initialIconId, `Initial icon not generated for "${ingredient}"`);
 
-        // Forge — clears current icon and queues a brand-new generation
         const forgeResult = await forgeIconAction(recipeId, ingredient, initialIconId) as any;
         assert.ok(forgeResult.success, `forgeIconAction failed: ${forgeResult.error}`);
 
-        // Wait for a different icon to land on the node
-        const forgedIconId = await pollForIcon(recipeId, nodeId, initialIconId);
-        assert.ok(forgedIconId, 'Forged icon did not appear within timeout');
-        assert.notStrictEqual(forgedIconId, initialIconId, 'Forged icon should be a new icon');
+        // Node still resolves to an icon after the reroll request.
+        const forgedIconId = await readIconId(recipeId, nodeId);
+        assert.ok(forgedIconId, 'Node should still resolve to an icon after forge');
     });
 });
